build(ci): new CI/CD pipeline capable of building only changed images
1. It can be triggered based on changed files, or manually.
- for PRs it builds all changed images relative to the target branch,
both directly changed images, and changes in the base images.
The dependencies are captured in a separate json file which is
interpreted by a dedicated python script.
- for branch builds it looks at the changes in the last commit.
The logic for dependencies is similar to PRs
- in case of manual parameters the users can build any images
this is useful in case of rebuilds or clean-building everything

2. Needed to make some changes to the makefiles, and split
the .mk files, in order to better track which variables
impact witch image build. Otherwise a change in a config.mk would
trigger all builds.

3. Deleted the folder for the missing/unneeded image busybox.

4. Tweaked the curl installation since it was failing against
newer rockly linux package versions.

5. Refactor building/applying rootfs diffs. This change requires a
stacker version based on a private branch, which allows importing
files straight under the / directory.
- we we have actual reuse of base layers between images instead of a
flat hierarchy
- this was also needed because images depended on build-only layers
defined in stackerfiles of other images, making partial builds
impossible since the build-only dependencies are not available

6. Start publishing the devel layer and add dependencies for cmake.

7. Update the build scripts to pull unchanged dependencies
from the registry used to publish images

8. Temporarily remove squashfs builds, as stacker is throwing an error.

Signed-off-by: Andrei Aaron <aaaron@luxoft.com>