refactor(pkgs): dynamically indentify rpms/debs package and versions to install
Use 2 different package lists to track rpms/debs to be installed:
- `${{DISTRO}}-pkgs.sh` - contains selections of unversioned packages we want or don't want installed in our image - this file is edited manually by the dev
- `${{DISTRO}}-pkgs` - contains all versioned packages that will be installed in the container image - this file is edited only by the dev running `make pkgs`, which is using `${{DISTRO}}-pkgs.sh` as input
The idea is that reusing the same unmodified `${{DISTRO}}-pkgs` file is supposed to keep track of what changed in the image and when, and also is supposed to guarantee the builds can be reproduced -> but this is not possible (yet?), since the content of the yum/apt repos is not under our control.
While using unversioned packages in `${{DISTRO}}-pkgs.sh` makes developer life easy when deciding what to install and what not to install.
This PR should also take care of fixing the repeated package downloads for building an image - https://github.com/project-stacker/c3/issues/87, and cut some build time as a result.

Other notable changes:
- include all tooling necessary to install packages in the build-env layer
- also make sure to import the entire folder containing build scripts
- add make target at repo level to update pkgs version
- fix downloading i686 dependencies
- openj9-devel image is now based on base-devel
- GH workflow to automatically commit changed files to main branch

Signed-off-by: Andrei Aaron <aaaron@luxoft.com>