feat(pkgs): support generating package lists for arm64
We need separate lists per architecture as the needed packages may change. For example in
base-devel (debian and ubuntu) we have some extra packages in one of the list wich are not in the other.

Update the build-env recipe to make the customization needed to download both amd64 and arm64 packages
We use a amd64 machine to download packages for all arches for several reasons:
- GH only provides amd64 runners (also we may need to update packages locally for testing)
- We want all the updates in the same workspace in order to create a single commit with all updates
in case someone updates packages upstream on the repositories

Update the python script to add a custom architecture parameter

Update the workflow generating the updated package lists with versions

New covered distros:
- Ubuntu arm64 (needs custom apt configuration)
- Debian arm64 (no need for custom apt configuration)
- Rocklinux aarch64 (no need for custom yum configuration)

Signed-off-by: Andrei Aaron <aaaron@luxoft.com>